merge

+++ mini-lsm/src/compact/leveled.rs
@@ -1,7 +1,10 @@
 use std::collections::HashSet;
 
+use serde::{Deserialize, Serialize};
+
 use crate::lsm_storage::LsmStorageState;
 
+#[derive(Serialize, Deserialize)]
 pub struct LeveledCompactionTask {
     // if upper_level is `None`, then it is L0 compaction
     pub upper_level: Option<usize>,
@@ -75,7 +78,7 @@ impl LeveledCompactionController {
                     .sum::<u64>() as usize,
             );
         }
-        let base_level_size_bytes = self.options.base_level_size_mb * 1024 * 1024;
+        let base_level_size_bytes = self.options.base_level_size_mb as usize * 1024 * 1024;
 
         // select base level and compute target level size
         target_level_size[self.options.max_levels - 1] =
